+++ src/domains/lightning/mod.rs
@@ -1,4 +1,4 @@
 pub mod adapters;
 pub mod api;
 pub mod entities;
-pub mod usecases;
+pub mod services;
