+++ src/domains/lightning/services/lightning/invoices.rs
@@ -7,7 +7,7 @@ use crate::{
     domains::{
         lightning::{
             entities::{LightningInvoice, LightningInvoiceDeleteFilter},
-            usecases::LightningInvoicesUseCases,
+            services::LightningInvoicesUseCases,
         },
         users::entities::{AuthUser, Permission},
     },
