+++ src/domains/lightning/services/lightning/lightning_service.rs
