+++ src/domains/lightning/services/lightning/mod.rs
@@ -1,0 +1,9 @@
+mod addresses;
+mod invoices;
+mod lightning_service;
+mod node;
+mod payments;
+mod payments_processor;
+
+pub use lightning_service::LightningService;
+pub use payments_processor::BreezPaymentsProcessor;
