+++ src/domains/lightning/services/mod.rs
@@ -1,0 +1,11 @@
+pub mod lightning;
+mod lightning_use_cases;
+mod payments_processor;
+pub mod wallet;
+mod wallet_use_cases;
+
+pub use lightning::LightningService;
+pub use lightning_use_cases::*;
+pub use payments_processor::PaymentsProcessorUseCases;
+pub use wallet::WalletService;
+pub use wallet_use_cases::WalletUseCases;
