+++ src/domains/lightning/services/wallet/mod.rs
@@ -1,0 +1,3 @@
+mod wallet_service;
+
+pub use wallet_service::WalletService;
