+++ src/domains/lightning/services/wallet_use_cases.rs
