--- src/domains/lightning/usecases/lightning_service/mod.rs
@@ -1,9 +1,0 @@
-mod addresses;
-mod breez_payments_processor;
-mod invoices;
-mod node;
-mod payments;
-mod service;
-
-pub use breez_payments_processor::BreezPaymentsProcessor;
-pub use service::LightningService;
