--- src/domains/lightning/usecases/mod.rs
@@ -1,9 +1,0 @@
-pub mod lightning_service;
-mod lightning_use_cases;
-pub mod wallet_service;
-mod wallet_use_cases;
-
-pub use lightning_service::LightningService;
-pub use lightning_use_cases::*;
-pub use wallet_service::WalletService;
-pub use wallet_use_cases::WalletUseCases;
