--- src/domains/lightning/usecases/wallet_service/mod.rs
@@ -1,4 +1,0 @@
-mod service;
-mod wallet;
-
-pub use service::WalletService;
