--- src/domains/lightning/usecases/wallet_service/service.rs
@@ -1,11 +1,0 @@
-use crate::domains::lightning::adapters::LightningRepository;
-
-pub struct WalletService {
-    pub store: Box<dyn LightningRepository>,
-}
-
-impl WalletService {
-    pub fn new(store: Box<dyn LightningRepository>) -> Self {
-        WalletService { store }
-    }
-}
